Update  - Move api check to index.js and require check pass to load modules  - Update api warning messages
