+++ includes/enqueue-scripts.php
@@ -64,6 +64,8 @@ function edac_admin_enqueue_scripts() {
 			
 			);
 	
+			$pro = edac_check_plugin_active( 'accessibility-checker-pro/accessibility-checker-pro.php' ) && EDAC_KEY_VALID;
+	
 			wp_enqueue_script( 'edac-editor-app', plugin_dir_url( __DIR__ ) . 'build/editorApp.bundle.js', false, EDAC_VERSION, false );
 	
 			wp_localize_script(
@@ -76,6 +78,7 @@ function edac_admin_enqueue_scripts() {
 					'edacApiUrl'  => esc_url_raw( rest_url() . 'accessibility-checker/v1' ),
 					'baseurl'     => plugin_dir_url( __DIR__ ),
 					'active'      => $active,
+					'pro'         => $pro,
 					'scanUrl'     => get_preview_post_link(
 						$post_id, 
 						array( 'edac_pageScanner' => 1 )
