+++ src/editorApp/settings.js
@@ -1,5 +1,5 @@
 export const settings = {
 	JS_SCAN_ENABLED: true,
 	INFO_ENABLED : true,
-	DEBUG_ENABLED: true
+	DEBUG_ENABLED: false
 };
